Restructure source code.
