+++ src/common/bipart_memory_stream.h
