+++ src/common/circular_buffer.cpp
@@ -1,4 +1,4 @@
-#include "circular_buffer.h"
+#include "common/circular_buffer.h"
 
 #include <algorithm>
 #include <cassert>
