+++ src/common/circular_buffer.h
