+++ src/common/in_memory_stream.h
