+++ src/lightstep_span.h
@@ -2,8 +2,8 @@
 
 #include "lightstep-tracer-common/collector.pb.h"
 #include "lightstep_span_context.h"
-#include "logger.h"
-#include "recorder.h"
+#include "common/logger.h"
+#include "recorder/recorder.h"
 
 #include <opentracing/span.h>
