+++ src/recorder/auto_recorder.h
@@ -8,9 +8,9 @@
 #include <thread>
 #include "condition_variable_wrapper.h"
 #include "lightstep-tracer-common/collector.pb.h"
-#include "logger.h"
-#include "recorder.h"
-#include "report_builder.h"
+#include "common/logger.h"
+#include "recorder/recorder.h"
+#include "recorder/report_builder.h"
 
 namespace lightstep {
 // AutoRecorder buffers spans finished by a tracer and sends them over to
