+++ src/recorder/manual_recorder.cpp
@@ -1,5 +1,5 @@
 #include "manual_recorder.h"
-#include "utility.h"
+#include "common/utility.h"
 
 namespace lightstep {
 //------------------------------------------------------------------------------
